+++ CHANGELOG.md
@@ -13,6 +13,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fix
 - Added isWindows() utility method to centralize OS detection and restrict system tray minimization feature to Windows only. Was causing crash if activated on Linux (Debian).
 
+### Feat
+- Added stream title in output filename
+
 ---
 
 ## [0.1.0] - 2025-09-15